pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "cloak-privacy"
version = "0.1.0"
description = "Privacy-first CLI to detect and scrub PII/PHI from data"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{name = "Your Name"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Topic :: Security",
  "Topic :: Text Processing :: Linguistic"
]
dependencies = [
  "typer>=0.12.5",
  "rich>=13.7.1",
  "pydantic>=2.8.2",
  "structlog>=24.1.0",
  "pyyaml>=6.0.2",
  "regex>=2024.7.24",
  "jinja2>=3.1.4",
  "fsspec>=2024.6.1"
]

[project.optional-dependencies]
s3 = ["s3fs>=2024.6.1"]
spacy = ["spacy>=3.7.5"]
hf = ["transformers>=4.43.3", "torch>=2.3.1 ; platform_system!='Darwin' or platform_machine!='arm64'"]
tui = ["textual>=0.62.0"]
dev = ["pytest>=8.2.1", "coverage>=7.6.1", "ruff>=0.5.5", "mypy>=1.10.0", "types-PyYAML"]

[project.scripts]
cloak = "cloak.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cloak"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E","F","I","UP","B"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
